Create CustomTour component
