+++ frontend/package.json
@@ -19,6 +19,7 @@
     "@types/react-router-dom": "^5.1.7",
     "@types/react-scroll": "^1.8.2",
     "@types/react-select": "^3.0.26",
+    "@types/reactour": "^1.18.1",
     "@types/styled-components": "^5.1.7",
     "async-lock": "^1.2.8",
     "axios": "^0.21.1",
